refactor(tests): move tests to colocate with relevant components
