+++ frontend/jest.config.js
@@ -2,7 +2,7 @@
 module.exports = {
 	preset: "ts-jest",
 	testEnvironment: "jsdom",
-	setupFilesAfterEnv: ["./tests/testSetup.ts"],
+	setupFilesAfterEnv: ["./src/tests/testSetup.ts"],
 	transform: {
 		"^.+\\.(ts|tsx)$": "ts-jest",
 	},
