+++ frontend/package.json
@@ -13,8 +13,8 @@
   "scripts": {
     "start": "parcel serve src/index.html --no-cache",
     "build": "parcel build src/index.html",
-    "lint": "rome check tests src *.js --verbose && rome format tests src *.js --verbose",
-    "lint:fix": "rome check tests src ./*.js --apply --verbose && rome format tests src ./*.js --write --verbose",
+    "lint": "rome check src *.js --verbose && rome format src *.js --verbose",
+    "lint:fix": "rome check src ./*.js --apply --verbose && rome format src ./*.js --write --verbose",
     "test": "yarn jest",
     "typecheck": "yarn tsc --noEmit"
   },
